Deleted package-lock.json, moved nested util file
